refactoring post merge of job
