--- src/Core/Analytics/Helpers/QuarterlySummary.cs
@@ -1,11 +1,0 @@
-﻿namespace Biobanks.Analytics.Helpers
-{
-    public class QuarterlySummary
-    {
-        public string Biobank { get; set; }
-        public string Quarter { get; set; }
-        public int Count { get; set; }
-        public int Total { get; set; }
-
-    }
-}
